use buildLocalPathToJail instead of direct JAILED_DOCUMENT_ROOT concat
e.g. download as pdf/save as pdf

move buildLocalPathToJail to FileUtil instead of JailUtil given
that the JailUtil code isn't built on mobile
